database/scripts/python/requirements.txt
# Travel Book JSON Generator - Python Dependencies
# Install with: uv pip install -r requirements.txt

# LLM API
groq>=0.4.0

# Data processing
pandas>=2.0.0

# Geocoding
geopy>=2.3.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Enhanced CLI output
rich>=13.0.0
